Add files via upload
The Carabao Crossing seamlessly blends the nostalgia of classic gaming with the colorful essence of Filipino culture, drawing inspiration from revered titles like "Frogger" and the beloved world of "Animal Crossing." In this engaging adventure, players aim to reach the water to score points, but with a unique twist: after ascending to the water from above, they must then navigate downward, avoiding obstacles like feces along the way. It's a test of skill, timing, and strategy set against a backdrop that celebrates the lively spirit of Filipino life, offering an immersive experience that pays homage to gaming classics while embracing the rich cultural heritage of the Philippines.